Inform of file encryption/decryption status
direncrypt didn't check or report status in gpgops module.
There are multiple causes why the process might fail, for example
the GPG key created only for signing. Both encrypt and decrypt
now check for status and report to stdout in case of issues.
An error in the process doesn't stop it, so the same error might
happen to all files.

Related issue:
https://github.com/dmarsic/direncrypt/issues/7