Move direncrypt modules directly under root directory
A long time ago when I was starting with Python it made sense to
have library modules under a directory named `lib`. It no longer
makes sense to me, and it's easier to maintain the project if
they use a simpler organisation. Therefore all files in `lib/direncrypt`
have been moved to `direncrypt`.

Related issue:
https://github.com/dmarsic/direncrypt/issues/3